restore environment
