+++ 02.Kernel64/Makefile_backup
@@ -1,0 +1,69 @@
+NASM64 = nasm
+GCC64 = gcc -c -ffreestanding
+LD64 = ld -melf_x86_64 -T ../elf_x86_64.x -nostdlib -e Main -Ttext 0x200000
+
+OBJCOPY64 = objcopy -j .text -j .data -j .rodata -j .bss -S -O binary
+
+# 디렉터리 정의
+OBJECTDIRECTORY = Temp
+SOURCEDIRECTORY	= Source
+
+all: prepare Kernel64.bin
+
+prepare:
+	mkdir -p $(OBJECTDIRECTORY)
+
+$(OBJECTDIRECTORY)/EntryPoint.bin : $(SOURCEDIRECTORY)/EntryPoint.s
+	$(NASM64) -f elf64 -o $@ $< 
+
+dep:
+	@echo === Make Dependancy File ===
+	make -C $(OBJECTDIRECTORY) -f ../Makefile InternalDependency
+	@echo === Dependancy Search Complete ===
+
+ExecuteInternalBuild: dep
+	make -C $(OBJECTDIRECTORY) -f ../Makefile Kernel64.elf
+
+#Kernel64.bin: ExecuteInternalBuild
+#	$(OBJCOPY64) $(OBJECTDIRECTORY)/Kernel64.elf $@
+
+$(OBJECTDIRECTORY)/Kernel64.elf.bin: ExecuteInternalBuild
+	$(OBJCOPY64) $(OBJECTDIRECTORY)/Kernel64.elf $@
+
+Kernel64.bin: $(OBJECTDIRECTORY)/EntryPoint.bin $(OBJECTDIRECTORY)/Kernel64.elf.bin
+	cat $^ > $@
+
+clean:
+	rm -f *.bin
+	rm -f $(OBJECTDIRECTORY)/*.*
+
+################################################################################
+# Make에 의해 다시 호출되는 부분, Temp 디렉터리를 기준으로 수행됨
+################################################################################
+# 빌드할 C 소스 파일 정의, Temp 디렉터리를 기준으로 설정
+ENTRYPOINTSOURCEFILE = ../$(SOURCEDIRECTORY)/EntryPoint.s
+ENTRYPOINTOBJECTFILE = EntryPoint.o
+CENTRYPOINTOBJECTFILE = Main.o
+CSOURCEFILES = $(wildcard ../$(SOURCEDIRECTORY)/*.c)
+ASSEMBLYSOURCEFILES = $(wildcard ../$(SOURCEDIRECTORY)/*.asm)
+COBJECTFILES = $(subst Main.o, , $(notdir $(patsubst %.c,%.o,$(CSOURCEFILES))))
+ASSEMBLYOBJECTFILES = $(notdir $(patsubst %.asm,%.o,$(ASSEMBLYSOURCEFILES)))
+
+$(ENTRYPOINTOBJECTFILE): $(ENTRYPOINTSOURCEFILE)
+	$(NASM64) -o $@ $<
+
+%.o: ../$(SOURCEDIRECTORY)/%.c
+	$(GCC64) -c $<
+
+%.o: ../$(SOURCEDIRECTORY)/%.asm
+	$(NASM64) -f elf64 -o $@ $<
+
+InternalDependency:
+	$(GCC64) -MM $(CSOURCEFILES) > Dependency.dep
+
+Kernel64.elf: $(CENTRYPOINTOBJECTFILE) $(COBJECTFILES) $(ASSEMBLYOBJECTFILES)
+	$(LD64) -o $@ $^
+
+ifeq (Dependency.dep, $(wildcard Dependency.dep))
+include Dependency.dep
+endif
