+++ 02.Kernel64/Source/Types.h
@@ -14,7 +14,6 @@
 // 매크로
 //
 ////////////////////////////////////////////////////////////////////////////////
-// 기본 타입 관련 매크로
 #define BYTE    unsigned char
 #define WORD    unsigned short
 #define DWORD   unsigned int
@@ -24,6 +23,7 @@
 #define TRUE    1
 #define FALSE   0
 #define NULL    0
+
 
 ////////////////////////////////////////////////////////////////////////////////
 //
